Rename Releaf::BaseController to Releaf::ActionController and extract modules.
