+++ releaf-core/lib/generators/dummy/templates/controllers/admin/authors_controller.rb
@@ -3,10 +3,7 @@ class Admin::AuthorsController < Releaf::ActionController
     [:name, :surname, books: [:title, chapters: [:title]]]
   end
 
-  protected
-
-  def setup
-    super
-    self.resources_per_page = params.has_key?(:show_all) ? nil : 20
+  def resources_per_page
+    params.has_key?(:show_all) ? nil : 20
   end
 end
